app/requirements.txt
streamlit>=1.24.0
pillow>=9.0.0
google-generativeai>=0.3.0